refactor(cas): mv to components
