+++ src/renderer/features/accounts-structure/components/AccountStructureNode.tsx
