+++ src/renderer/features/accounts-structure/components/AccountsStructure.tsx
