+++ src/renderer/features/accounts-structure/components/AccountsStructureModal.tsx
